requirements.txt
functions-framework==3.*
google-cloud-pubsub
google-cloud-logging